ingestion-service/requirements.txt
fastapi
uvicorn
pydantic
python-json-logger
kafka-python